make bitmap nicer

+++ src/table/column/utf8/mod.rs
@@ -66,7 +66,7 @@ impl<'a> Column for Utf8Column<'a> {
                         )));
                     }
 
-                    if !seg.nulls.is_valid(*segment_offset) {
+                    if !seg.nulls.is_valid(*segment_offset as u64) {
                         builder.append_null();
                     } else {
                         let (start, end) = seg.string_range(*segment_offset);
@@ -88,7 +88,7 @@ impl<'a> Column for Utf8Column<'a> {
 
         for seg in &self.segments {
             for i in 0..seg.header.num_values {
-                if !seg.nulls.is_valid(i) {
+                if !seg.nulls.is_valid(i as u64) {
                     builder.append_null();
                 } else {
                     let (start, end) = seg.string_range(i);
